refactor: rename `Nazuki.CodeGen` to `Nazuki.VM`
