+++ src/Nazuki/VM/Arch232/IO.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE BlockArguments #-}
 {-# LANGUAGE OverloadedStrings #-}
 
-module Nazuki.CodeGen.Arch232.IO
+module Nazuki.VM.Arch232.IO
   ( doGetc,
     doPutc,
   )
@@ -9,9 +9,9 @@ where
 
 import Control.Monad
 import qualified Data.Bits as Bits
-import Nazuki.CodeGen.Arch232.Common
-import Nazuki.CodeGen.Core
-import Nazuki.CodeGen.Util
+import Nazuki.VM.Arch232.Common
+import Nazuki.VM.Core
+import Nazuki.VM.Util
 
 doGetc :: Oper
 doGetc = do
