+++ test/Nazuki/VM/Arch2561/ByteSpec.hs
@@ -1,12 +1,12 @@
 {-# LANGUAGE BlockArguments #-}
 {-# LANGUAGE OverloadedStrings #-}
 
-module Nazuki.CodeGen.Arch2561.ByteSpec (spec) where
+module Nazuki.VM.Arch2561.ByteSpec (spec) where
 
 import qualified Data.Text as T
-import Nazuki.CodeGen.Arch2561.Byte
-import Nazuki.CodeGen.Core (generate)
 import Nazuki.Runtime.Runner (run)
+import Nazuki.VM.Arch2561.Byte
+import Nazuki.VM.Core (generate)
 import Test.Hspec
 
 spec :: Spec
